requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==4.2.13
django-allauth==0.63.5
django-appconf==1.0.6
django-image-cropping==1.7
django-phonenumber-field==8.0.0
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
fpdf==1.7.2
git-filter-repo==2.45.0
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
oauthlib==3.2.2
openpyxl==3.1.5
phonenumbers==8.13.39
pillow==10.3.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-decouple==3.8
python3-openid==3.2.0
razorpay==1.4.2
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
